Updating documentation
