+++ scripts/nb_counts.csv
